Update README.md. Better dev environment.
Remove the Public API Keys and config from the code and pull it from the dynamic reserved URLs from firebase.  In doing that, we have to also run the firebase dev server and proxy unknown requests (i.e. /__/firebase/*) to it so they can be served.  In production, Firebase itself will handle this.